chore(graphql-schema): rename graphql-scheme to graphql-schema
