+++ lib/ace/mode/graphqlschema.js
@@ -1,7 +1,7 @@
 /* ***** BEGIN LICENSE BLOCK *****
  * Distributed under the BSD license:
  *
- * Copyright (c) 2010, Ajax.org B.V.
+ * Copyright (c) 2012, Ajax.org B.V.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -32,43 +32,20 @@ define(function(require, exports, module) {
 "use strict";
 
 var oop = require("../lib/oop");
-var TextHighlightRules = require("./text_highlight_rules").TextHighlightRules;
+var TextMode = require("./text").Mode;
+var GraphQLSchemaHighlightRules = require("./graphqlschema_highlight_rules").GraphQLSchemaHighlightRules;
+var FoldMode = require("./folding/cstyle").FoldMode;
 
-var GraphQLSchemeHighlightRules = function() {
-
-    var keywords = (
-      "type|interface|union|enum|schema|input|implements|extends|scalar"
-    );
-
-    var dataTypes = (
-      "Int|Float|String|ID|Boolean"
-    );
-
-    var keywordMapper = this.createKeywordMapper({
-        "keyword": keywords,
-        "storage.type": dataTypes
-    }, "identifier");
-
-    this.$rules = {
-      "start" : [ {
-        token : "comment",
-        regex : "#.*$"
-      }, {
-        token : "paren.lparen",
-        regex : /[\[({]/,
-        next  : "start"
-      }, {
-        token : "paren.rparen",
-        regex : /[\])}]/
-      }, {
-        token : keywordMapper,
-        regex : "[a-zA-Z_$][a-zA-Z0-9_$]*\\b"
-      } ]
-    };
-    this.normalizeRules();
+var Mode = function() {
+    this.HighlightRules = GraphQLSchemaHighlightRules;
+    this.foldingRules = new FoldMode();
 };
+oop.inherits(Mode, TextMode);
 
-oop.inherits(GraphQLSchemeHighlightRules, TextHighlightRules);
+(function() {
+    this.lineCommentStart = "#";
+    this.$id = "ace/mode/graphqlschema";
+}).call(Mode.prototype);
 
-exports.GraphQLSchemeHighlightRules = GraphQLSchemeHighlightRules;
+exports.Mode = Mode;
 });
