+++ lib/ace/mode/graphqlschema_highlight_rules.js
@@ -1,7 +1,7 @@
 /* ***** BEGIN LICENSE BLOCK *****
  * Distributed under the BSD license:
  *
- * Copyright (c) 2012, Ajax.org B.V.
+ * Copyright (c) 2010, Ajax.org B.V.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -32,20 +32,43 @@ define(function(require, exports, module) {
 "use strict";
 
 var oop = require("../lib/oop");
-var TextMode = require("./text").Mode;
-var GraphQLSchemeHighlightRules = require("./graphqlscheme_highlight_rules").GraphQLSchemeHighlightRules;
-var FoldMode = require("./folding/cstyle").FoldMode;
+var TextHighlightRules = require("./text_highlight_rules").TextHighlightRules;
 
-var Mode = function() {
-    this.HighlightRules = GraphQLSchemeHighlightRules;
-    this.foldingRules = new FoldMode();
+var GraphQLSchemaHighlightRules = function() {
+
+    var keywords = (
+      "type|interface|union|enum|schema|input|implements|extends|scalar"
+    );
+
+    var dataTypes = (
+      "Int|Float|String|ID|Boolean"
+    );
+
+    var keywordMapper = this.createKeywordMapper({
+        "keyword": keywords,
+        "storage.type": dataTypes
+    }, "identifier");
+
+    this.$rules = {
+      "start" : [ {
+        token : "comment",
+        regex : "#.*$"
+      }, {
+        token : "paren.lparen",
+        regex : /[\[({]/,
+        next  : "start"
+      }, {
+        token : "paren.rparen",
+        regex : /[\])}]/
+      }, {
+        token : keywordMapper,
+        regex : "[a-zA-Z_$][a-zA-Z0-9_$]*\\b"
+      } ]
+    };
+    this.normalizeRules();
 };
-oop.inherits(Mode, TextMode);
 
-(function() {
-    this.lineCommentStart = "#";
-    this.$id = "ace/mode/graphqlscheme";
-}).call(Mode.prototype);
+oop.inherits(GraphQLSchemaHighlightRules, TextHighlightRules);
 
-exports.Mode = Mode;
+exports.GraphQLSchemaHighlightRules = GraphQLSchemaHighlightRules;
 });
