+++ lib/ace/snippets/graphqlschema.js
@@ -1,0 +1,7 @@
+define(function(require, exports, module) {
+"use strict";
+
+exports.snippetText = require("../requirejs/text!./graphqlschema.snippets");
+exports.scope = "graphqlschema";
+
+});
