+++ lib/ace/snippets/graphqlschema.snippets
