--- lib/ace/snippets/graphqlscheme.js
@@ -1,7 +1,0 @@
-define(function(require, exports, module) {
-"use strict";
-
-exports.snippetText = require("../requirejs/text!./graphqlscheme.snippets");
-exports.scope = "graphqlscheme";
-
-});
